test(runner): split slow runner test files
The runner test job assigns files to shards by sorted round-robin. A few large test files kept much of the work in one shard, which made CI runtime uneven even when the file counts were balanced.

Split the largest runner test files into smaller files that preserve the existing coverage while giving the shard selector more units to distribute. The engine tests now follow their existing top-level suites and share common setup through engine-test-support.ts. The fetchData mutex coverage is divided into core mutex behavior, reactive fetch state, and protected-request auth files. The profile space collision coverage now has one assertion per test file, with shared profile creation code in a support module.

The profile assertions still check the same behavior, but each file now creates only the spaces that assertion needs instead of running all profile setup before every assertion.
